extract cli and public api modules
