+++ tests/test_serialization.py
@@ -6,6 +6,7 @@ import json
 from collections import defaultdict
 
 import gem
+import gem.api
 from gem.results.models import ParsedMatch, ParsedPlayer
 
 
@@ -45,7 +46,7 @@ class TestSerializationHelpers:
     def test_parse_to_json_uses_parse_result(self, monkeypatch):
         fake_match = ParsedMatch(match_id=999)
 
-        monkeypatch.setattr(gem, "parse", lambda path: fake_match)
+        monkeypatch.setattr(gem.api, "parse", lambda path: fake_match)
 
         payload = gem.parse_to_json("dummy.dem")
         decoded = json.loads(payload)
